extension: split content-id resolution out of perfcache.c into content_resolve.c
word_segment (UAX#29), resolve_phrase (longest witnessed span), and realize
(best surface for an id) are SPI-backed reads over witnessed/consensus data — a
different layer than the T0 id<->codepoint floor. perfcache.c stays the floor.
Already built + deployed (build-win-ext); committing to baseline the racers share.
