+++ extension/laplace_substrate/src/perfcache.c
@@ -11,14 +11,14 @@
  * codepoint_for_id(bytea) is the reverse T0 lookup that retired the 1.1M-row
  * codepoint_render table: a binary search over a per-backend index of
  * codepoints sorted by record hash, built once on first use (~4.4 MB).
+ *
+ * Content-id RESOLUTION surfaces (word_segment, resolve_phrase, realize) that
+ * read witnessed/consensus data via SPI live in content_resolve.c, not here.
  */
 #include "postgres.h"
 
 #include "fmgr.h"
-#include "funcapi.h"
-#include "catalog/pg_collation.h"
 #include "utils/builtins.h"
-#include "utils/formatting.h"
 #include "utils/guc.h"
 #include "utils/memutils.h"
 
@@ -27,8 +27,6 @@
 #include "laplace/core/hash128.h"
 
 #include "perfcache_native.h"
-#include "spi_common.h"
-#include "spi_nested.h"
 
 static char *perfcache_path = NULL;
 
@@ -216,314 +214,4 @@ pg_laplace_is_all_whitespace(PG_FUNCTION_ARGS)
     t = PG_GETARG_TEXT_PP(0);
     PG_RETURN_BOOL(laplace_text_is_all_whitespace(
         (const uint8_t *) VARDATA_ANY(t), VARSIZE_ANY_EXHDR(t)) != 0);
-}
-
-/* ---- omniglottal word segmentation (engine UAX#29) --------------------- */
-
-/* word_segment(text) -> (ord, word, id): the engine's UAX#29 word law as a SRF,
- * replacing the SQL `regexp_split_to_array(p_text, '[^[:alnum:]''-]+')` token
- * fork in prompt_words/resolve_phrase. Same decomposer word_id/deposition use,
- * so 中文 segments by the same law as English, and each word carries its
- * deposition-matching id — the exact-case probe needs no second word_id(). */
-
-typedef struct
-{
-    ReturnSetInfo *rsinfo;
-} word_seg_ctx;
-
-static void
-word_seg_emit(void *ctx_, uint32_t ordinal,
-              const uint8_t *word_utf8, uint32_t word_len,
-              const hash128_t *id)
-{
-    word_seg_ctx *ctx = (word_seg_ctx *) ctx_;
-    Datum         values[3];
-    bool          nulls[3] = { false, false, false };
-
-    values[0] = Int32GetDatum((int32) ordinal);
-    values[1] = PointerGetDatum(
-        cstring_to_text_with_len((const char *) word_utf8, (int) word_len));
-    values[2] = hash128_to_datum(id);
-    tuplestore_putvalues(ctx->rsinfo->setResult, ctx->rsinfo->setDesc, values, nulls);
-}
-
-PG_FUNCTION_INFO_V1(pg_laplace_word_segment);
-
-Datum
-pg_laplace_word_segment(PG_FUNCTION_ARGS)
-{
-    text         *t;
-    word_seg_ctx  ctx;
-    int           rc;
-
-    InitMaterializedSRF(fcinfo, 0);
-    if (PG_ARGISNULL(0))
-        return (Datum) 0;
-    t = PG_GETARG_TEXT_PP(0);
-    if (VARSIZE_ANY_EXHDR(t) == 0)
-        return (Datum) 0;
-    if (!laplace_perfcache_ready())
-        ereport(ERROR,
-                (errcode(ERRCODE_OBJECT_NOT_IN_PREREQUISITE_STATE),
-                 errmsg("word_segment requires the T0 perfcache")));
-
-    ctx.rsinfo = (ReturnSetInfo *) fcinfo->resultinfo;
-    rc = laplace_content_word_segment((const uint8_t *) VARDATA_ANY(t),
-                                      (size_t) VARSIZE_ANY_EXHDR(t),
-                                      word_seg_emit, &ctx);
-    if (rc != 0)
-        ereport(ERROR,
-                (errcode(ERRCODE_INVALID_PARAMETER_VALUE),
-                 errmsg("word_segment: segmentation failed (rc=%d)", rc)));
-    return (Datum) 0;
-}
-
-/* ---- longest witnessed phrase span (resolve_phrase) -------------------- */
-
-/* resolve_phrase(text) -> id: the longest contiguous word span (leftmost on
- * ties) whose composed content id is a witnessed entity. The old body built
- * EVERY span with an O(n^2) token self-join and called word_id() twice per
- * span; here one engine segmentation feeds a longest-first scan, each candidate
- * hashed once and probed once. Tries the surface span and its case-folded form
- * (the old body resolved purely case-folded). */
-
-typedef struct
-{
-    const uint8_t *base;
-    uint32_t      *off;
-    uint32_t      *len;
-    int            n;
-    int            cap;
-} phrase_ctx;
-
-static void
-phrase_collect_emit(void *ctx_, uint32_t ordinal,
-                    const uint8_t *word_utf8, uint32_t word_len,
-                    const hash128_t *id)
-{
-    phrase_ctx *ctx = (phrase_ctx *) ctx_;
-
-    (void) ordinal;
-    (void) id;
-    if (ctx->n == ctx->cap)
-    {
-        int newcap = ctx->cap ? ctx->cap * 2 : 16;
-
-        if (ctx->off == NULL)
-        {
-            ctx->off = (uint32_t *) palloc(sizeof(uint32_t) * newcap);
-            ctx->len = (uint32_t *) palloc(sizeof(uint32_t) * newcap);
-        }
-        else
-        {
-            ctx->off = (uint32_t *) repalloc(ctx->off, sizeof(uint32_t) * newcap);
-            ctx->len = (uint32_t *) repalloc(ctx->len, sizeof(uint32_t) * newcap);
-        }
-        ctx->cap = newcap;
-    }
-    /* store the byte offset, not the transient pointer (valid only during call) */
-    ctx->off[ctx->n] = (uint32_t) (word_utf8 - ctx->base);
-    ctx->len[ctx->n] = word_len;
-    ctx->n++;
-}
-
-static bool
-phrase_id_is_entity(hash128_t *id)
-{
-    Oid   argtypes[1] = { BYTEAOID };
-    Datum args[1];
-    int   rc;
-
-    args[0] = hash128_to_datum(id);
-    rc = SPI_execute_with_args(
-        "SELECT 1 FROM laplace.entities WHERE id = $1",
-        1, argtypes, args, NULL, true, 1);
-    return rc == SPI_OK_SELECT && SPI_processed > 0;
-}
-
-PG_FUNCTION_INFO_V1(pg_laplace_resolve_phrase);
-
-Datum
-pg_laplace_resolve_phrase(PG_FUNCTION_ARGS)
-{
-    text          *t;
-    const uint8_t *base;
-    phrase_ctx     ctx;
-    int            rc;
-    bool           spi_top = false;
-    bool           found = false;
-    hash128_t      found_id = { 0, 0 };
-
-    if (PG_ARGISNULL(0))
-        PG_RETURN_NULL();
-    t = PG_GETARG_TEXT_PP(0);
-    if (VARSIZE_ANY_EXHDR(t) == 0)
-        PG_RETURN_NULL();
-    if (!laplace_perfcache_ready())
-        ereport(ERROR,
-                (errcode(ERRCODE_OBJECT_NOT_IN_PREREQUISITE_STATE),
-                 errmsg("resolve_phrase requires the T0 perfcache")));
-
-    base = (const uint8_t *) VARDATA_ANY(t);
-    memset(&ctx, 0, sizeof(ctx));
-    ctx.base = base;
-    rc = laplace_content_word_segment(base, (size_t) VARSIZE_ANY_EXHDR(t),
-                                      phrase_collect_emit, &ctx);
-    if (rc != 0)
-        ereport(ERROR,
-                (errcode(ERRCODE_INVALID_PARAMETER_VALUE),
-                 errmsg("resolve_phrase: segmentation failed (rc=%d)", rc)));
-    if (ctx.n == 0)
-        PG_RETURN_NULL();
-
-    if (laplace_spi_connect(&spi_top) != SPI_OK_CONNECT)
-        elog(ERROR, "resolve_phrase: SPI_connect failed");
-
-    /* longest-first, leftmost on ties; the id is a stack value so it survives
-     * SPI_finish without a memory-context dance. */
-    for (int L = ctx.n; L >= 1 && !found; L--)
-    {
-        for (int i = 0; i + L <= ctx.n && !found; i++)
-        {
-            const uint8_t *sp = base + ctx.off[i];
-            size_t  splen = (size_t) ((ctx.off[i + L - 1] + ctx.len[i + L - 1])
-                                      - ctx.off[i]);
-            hash128_t id;
-            char     *lowered;
-
-            if (laplace_content_root_id(sp, splen, &id) == 0
-                && phrase_id_is_entity(&id))
-            {
-                found_id = id;
-                found = true;
-                break;
-            }
-            lowered = str_tolower((const char *) sp, splen, DEFAULT_COLLATION_OID);
-            if (lowered != NULL
-                && laplace_content_root_id((const uint8_t *) lowered,
-                                           strlen(lowered), &id) == 0
-                && phrase_id_is_entity(&id))
-            {
-                found_id = id;
-                found = true;
-                break;
-            }
-        }
-    }
-
-    laplace_spi_finish(spi_top);
-    if (!found)
-        PG_RETURN_NULL();
-    PG_RETURN_DATUM(hash128_to_datum(&found_id));
-}
-
-/* ---- realize: best surface text for an id (native COALESCE ladder) ------ */
-
-/* Run one realize branch; return its first column copied into the caller's
- * context (SPI_palloc survives SPI_finish), or NULL on no-row / NULL / empty.
- * The branch SQL is the SAME predicate the SQL COALESCE used, so semantics are
- * identical — only the short-circuit orchestration moves into C. */
-static text *
-realize_branch(const char *sql, Datum id, Datum lang, int nargs)
-{
-    Oid   argtypes[2] = { BYTEAOID, BYTEAOID };
-    Datum args[2] = { id, lang };
-    char  nulls[3] = "   ";
-    bool  isnull;
-    int   rc;
-    Datum d;
-    text *src;
-    text *dst;
-    Size  sz;
-
-    if (nargs == 2 && lang == (Datum) 0)
-        nulls[1] = 'n';
-    rc = SPI_execute_with_args(sql, nargs, argtypes, args, nulls, true, 1);
-    if (rc != SPI_OK_SELECT || SPI_processed == 0)
-        return NULL;
-    d = SPI_getbinval(SPI_tuptable->vals[0], SPI_tuptable->tupdesc, 1, &isnull);
-    if (isnull)
-        return NULL;
-    src = DatumGetTextPP(d);
-    if (VARSIZE_ANY_EXHDR(src) == 0)
-        return NULL;                       /* empty string is not a realization */
-    sz = VARSIZE_ANY(src);
-    dst = (text *) SPI_palloc(sz);         /* upper context: outlives SPI_finish */
-    memcpy(dst, src, sz);
-    return dst;
-}
-
-PG_FUNCTION_INFO_V1(pg_laplace_realize);
-
-Datum
-pg_laplace_realize(PG_FUNCTION_ARGS)
-{
-    Datum id, lang;
-    bool  spi_top = false;
-    text *out;
-
-    if (PG_ARGISNULL(0))
-        PG_RETURN_NULL();
-    id = PG_GETARG_DATUM(0);
-    lang = PG_ARGISNULL(1) ? (Datum) 0 : PG_GETARG_DATUM(1);
-
-    if (laplace_spi_connect(&spi_top) != SPI_OK_CONNECT)
-        elog(ERROR, "realize: SPI_connect failed");
-
-    out = realize_branch("SELECT NULLIF(laplace.render_text($1), '')", id, lang, 1);
-    if (out == NULL)
-        out = realize_branch(
-            /* render_text ONCE per candidate row (was SELECT + WHERE = 2x) */
-            "SELECT q.s FROM ("
-            "  SELECT laplace.render_text(hs.subject_id) AS s, "
-            "         (lang.object_id IS NOT NULL) AS lp, "
-            "         laplace.eff_mu(hs.rating, hs.rd) AS mu "
-            "  FROM laplace.consensus io "
-            "  JOIN laplace.consensus hs ON hs.object_id = io.subject_id "
-            "    AND hs.type_id = laplace.relation_type_id('HAS_SENSE') "
-            "  LEFT JOIN laplace.consensus lang ON lang.subject_id = hs.subject_id "
-            "    AND lang.type_id = laplace.relation_type_id('HAS_LANGUAGE') "
-            "    AND lang.object_id = $2 "
-            "  WHERE io.object_id = $1 "
-            "    AND io.type_id = laplace.relation_type_id('IS_SENSE_OF') "
-            "    AND NOT laplace.refuted(io.rating, io.rd)"
-            ") q WHERE q.s IS NOT NULL AND q.s <> '' "
-            "ORDER BY q.lp DESC, q.mu DESC LIMIT 1", id, lang, 2);
-    if (out == NULL)
-        out = realize_branch(
-            /* render_text ONCE per candidate row (was SELECT + WHERE = 2x) */
-            "SELECT q.s FROM ("
-            "  SELECT laplace.render_text(m.object_id) AS s, "
-            "         (lang.object_id IS NOT NULL) AS lp, "
-            "         laplace.eff_mu(m.rating, m.rd) AS mu "
-            "  FROM laplace.consensus m "
-            "  LEFT JOIN laplace.consensus lang ON lang.subject_id = m.object_id "
-            "    AND lang.type_id = laplace.relation_type_id('HAS_LANGUAGE') "
-            "    AND lang.object_id = $2 "
-            "  WHERE m.subject_id = $1 "
-            "    AND m.type_id = laplace.relation_type_id('IS_TRANSLATION_OF') "
-            "    AND NOT laplace.refuted(m.rating, m.rd)"
-            ") q WHERE q.s IS NOT NULL AND q.s <> '' "
-            "ORDER BY q.lp DESC, q.mu DESC LIMIT 1", id, lang, 2);
-    if (out == NULL)
-        out = realize_branch(
-            "SELECT regexp_replace(n.name, '^substrate/[a-z_]+/(.+)/v1$', '\\1') "
-            "FROM laplace.canonical_names n "
-            "WHERE n.id = $1 AND n.name LIKE 'substrate/%'", id, lang, 1);
-    if (out == NULL)
-        out = realize_branch(
-            "SELECT laplace.render_text(g.object_id) "
-            "FROM laplace.consensus g "
-            "WHERE g.subject_id = $1 AND g.type_id = laplace.relation_type_id('DEFINES') "
-            "  AND NOT laplace.refuted(g.rating, g.rd) "
-            "ORDER BY laplace.eff_mu(g.rating, g.rd) DESC "
-            "LIMIT 1", id, lang, 1);
-    if (out == NULL)
-        out = realize_branch("SELECT laplace.label($1)", id, lang, 1);
-
-    laplace_spi_finish(spi_top);
-
-    if (out == NULL)
-        PG_RETURN_NULL();
-    PG_RETURN_TEXT_P(out);
 }
